feat: add component Home
