+++ src/index.scss
@@ -40,3 +40,8 @@ a {
 a:hover {
   color: #535bf2;
 }
+.active {
+  border-bottom: 2px solid #646cff;
+  font-weight: bold;
+  color: #646cff !important;
+}
